+++ tests/inputs/if-comments-2.lua
@@ -1,0 +1,14 @@
+-- https://github.com/JohnnyMorganz/StyLua/issues/254
+if condition1 then
+	print("Do something")
+
+--[[
+	my multiline comment
+]]
+elseif condition2 then
+	print("Do something else")
+
+-- my single line comment
+elseif condition3 then
+	print("Do some final thing")
+end